refactor(appearance): move logo mark widget
